Completely redid the shared library approach. I've ditched the Redhat way since the original makefiles are good to go with only small tweaks.
The major difference will be that the shared libraries will have an soname
which explicitly spells out the OpenSSL version
ie. instead of libssl.so.3 we'll use libssl.so.0.9.7.3

No problems with having to identify if a newer version is binary compatible
and should be given the same soname.

*All* packages which use a shared libssl should have a depend on the
SBossl<ver>lib package.

Different SBossl<ver>lib packages can be installed at the same time since
it will *only* contain the shared libraries.