gcc: update gcc_compat function
To avoid dragging around a binary datafile in the tree the sqlite DB
has been converted to ascii. This requires loading it on each use but
the overhead should not be too bad and it's not performance critical.

The startup sanity checking was moved to a separate function,
gcc_compat_check which can be called before using gcc_compat.